Add windows builds
Add windows build and fix all related build failures. This currently only builds
`msvc-15` due to LLVM issues. Reduced build size by removing incremental linking
option, switching submodules to do shallow clones and removing unneeded LLVM
targets.

+++ Overlays/Windows/Overlays/D3D12/Runtime/WNGraphics/inc/Internal/D3D12/WNDevice.h
@@ -264,7 +264,7 @@ protected:
   bool initialize_buffer(buffer* _buffer, const size_t _size,
       const resource_states _usage) WN_GRAPHICS_OVERRIDE_FINAL;
   bool bind_buffer(buffer* _buffer, arena* _arena,
-      const size_t _offset) WN_GRAPHICS_OVERRIDE_FINAL;
+      const uint64_t _offset) WN_GRAPHICS_OVERRIDE_FINAL;
   void* map_buffer(buffer* _buffer) WN_GRAPHICS_OVERRIDE_FINAL;
   void unmap_buffer(buffer* _buffer) WN_GRAPHICS_OVERRIDE_FINAL;
   void destroy_buffer(buffer* _buffer) WN_GRAPHICS_OVERRIDE_FINAL;
